requirements.txt
fastapi[standard]
langchain
langchain-openai
docling